Database added

+++ app/src/main/java/hr/ferit/sandroblavicki/rma_lv3_zad2/listeners/PersonInteractionListener.kt
@@ -1,7 +1,7 @@
 package hr.ferit.sandroblavicki.rma_lv3_zad2.listeners
 
 interface PersonInteractionListener {
-    fun onRemove(id: Int)
-    fun onShowQuote(id: Int)
+    fun onRemove(id: Long)
+    fun onShowQuote(id: Long)
     fun onEditDetails(id: Int)
 }
